+++ src/impl/file_io.h
@@ -4,7 +4,6 @@
 #include <sys/types.h>
 
 #include <stdbool.h>
-#include <stdint.h>
 #include <stdio.h>
 
 #pragma GCC diagnostic push
@@ -24,6 +23,6 @@ void file_close(struct file*);
 
 off_t file_offset(const struct file*);
 
-ssize_t file_splice(struct file* file, int fd, uint64_t offset, uint64_t count);
+ssize_t file_splice(struct file* file, int fd, loff_t offset, size_t count);
 
 #endif
